+++ core/src/main/java/com/graphhopper/util/EdgeIteratorState.java
@@ -160,12 +160,11 @@ public interface EdgeIteratorState {
     EdgeIteratorState setDistance(double dist);
 
     /**
-     * @return the distance of the current edge in millimeters as a whole-number double.
-     * Whole-number doubles are exact up to 2^53, so summing mm values is lossless.
+     * @return the distance of the current edge in millimeters
      */
-    double getDistance_mm();
+    long getDistance_mm();
 
-    EdgeIteratorState setDistance_mm(double distance_mm);
+    EdgeIteratorState setDistance_mm(long distance_mm);
 
     /**
      * Returns edge properties stored in direction of the raw database layout. So do not use it directly, instead
